+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+* Added Message model structures
+* Added Suggestions to Repository
+* Moved Repository interface to root
+* Added JsonSerializable to Suggestion interface
 * Improved default isReady() implementation
 
 ## v0.2.1 (2024-05-20)